+++ lending/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.7.25",
-  genVersion: "2.481.0",
-  userAgent: "speakeasy-sdk/typescript 0.7.25 2.481.0 1.0.0 ryan-lending",
+  sdkVersion: "0.7.26",
+  genVersion: "2.483.1",
+  userAgent: "speakeasy-sdk/typescript 0.7.26 2.483.1 1.0.0 ryan-lending",
 } as const;